+++ _content/tour/algorithms-bits-seven.article
@@ -6,4 +6,4 @@ This section provides examples that perform bit manipulation.
 This sample program shows you how to check if an integer is even or odd
 using bit manipulation.
 
-.play algorithms/bits/iseven/iseven.go
+.play algorithms/bits/iseven.go
