+++ _content/tour/algorithms-data.article
@@ -1,7 +1,7 @@
 Data Structures
 This section provides data struct examples.
 
-* Hash Table
+* Hash Map
 
 This sample program implements a basic hash table.
 
@@ -37,7 +37,7 @@ This sample program implements a basic hash table.
        │ key │ value │
        └─────────────┘
 
-.play algorithms/data/hash/hash.go
+.play algorithms/data/hash_map.go
 
 * Linked List
 
@@ -62,7 +62,7 @@ This sample program implements a basic double linked list.
                        │ First │  Last │
                        └───────────────┘
 
-.play algorithms/data/list/list.go
+.play algorithms/data/list.go
 
 * Queue
 
@@ -80,7 +80,7 @@ This sample program implements a basic circular queue.
     └─────┘    | └─────┘    └─────┘    └─────┘    └─────┘ |    └─────┘
                └──────────────────────────────────────────┘
 
-.play algorithms/data/queue/queue.go
+.play algorithms/data/queue_circular.go
 
 * Stack
 
@@ -109,7 +109,7 @@ This sample program implements a basic stack.
         │  └─────┘  │
         └───────────┘
 
-.play algorithms/data/stack/stack.go
+.play algorithms/data/stack.go
 
 * Binary Tree
 
@@ -140,4 +140,4 @@ This sample program implements a basic binary tree.
        │ 35 │  │ 65 │  │ 78 │  │ 95 │
        └────┘  └────┘  └────┘  └────┘
 
-.play algorithms/data/tree/binary/binary.go
+.play algorithms/data/tree_binary.go
