+++ _content/tour/algorithms-strings.article
@@ -1,12 +1,64 @@
-Strings
-TODO - Strings
+String operations
+This section provides examples that perform string manipulation.
 
-* Strings
+* Permutation
 
-** Palindrome
+The sample program implements a check to see if a string is a
+permutation or not.
 
-.play algorithms/strings/palindrome/palindrome.go
+- See more at [[https://en.wikipedia.org/wiki/Permutation][https://en.wikipedia.org/wiki/Permutation]]
 
-** Permutation
+*Diagram*
 
-.play algorithms/strings/permutation/permutation.go
+    The permutation of a string is the set of all strings that contains the
+    same characters without the order of the arrangement of the characters
+    mattering.
+
+    ┌───┐┌───┐┌───┐
+    │ G ││ O ││ D │ ──────┐
+    └───┘└───┘└───┘       │ ────▷ Permutation
+    ┌───┐┌───┐┌───┐       │
+    │ D ││ O ││ G │ ──────┘
+    └───┘└───┘└───┘
+
+    ┌───┐┌───┐┌───┐
+    │ G ││ O ││ D │ ──────┐
+    └───┘└───┘└───┘       │ ────▷ NO
+    ┌───┐┌───┐            │
+    │ D ││ O │      ──────┘
+    └───┘└───┘
+
+    ┌───┐┌───┐┌───┐
+    │ 1 ││ 0 ││ 0 │ ──────┐
+    └───┘└───┘└───┘       │ ────▷ Permutation
+    ┌───┐┌───┐┌───┐       │
+    │ 0 ││ 0 ││ 1 │ ──────┘
+    └───┘└───┘└───┘
+
+.play algorithms/strings/permutation.go
+
+* Palindrome
+
+The sample program implements a check to see if a string is a
+palindrome or not.
+
+- See more at [[https://en.wikipedia.org/wiki/Palindrome][https://en.wikipedia.org/wiki/Palindrome]]
+
+*Diagram*
+
+    A palindrome is a word, number, phrase, or other sequence of symbols that
+    reads the same backwards as forwards.
+
+    ┌───┐┌───┐┌───┐
+    │ B ││ O ││ B │ ────▷ Palindrome
+    └───┘└───┘└───┘
+
+    ┌───┐┌───┐┌───┐
+    │ T ││ E ││ D │ ────▷ NO
+    └───┘└───┘└───┘
+
+    ┌───┐
+    │ G │           ────▷ Palindrome
+    └───┘
+
+.play algorithms/strings/palindrome.go
