+++ _content/tour/algorithms/bits/iseven.go
