+++ _content/tour/algorithms/data/hash_map.go
@@ -4,7 +4,7 @@
 // All material is licensed under the Apache License Version 2.0, January 2004
 // http://www.apache.org/licenses/LICENSE-2.0
 
-// A sample that program implements a basic hash table.
+// This sample program shows you how to write a basic hash table.
 package main
 
 import (
