+++ _content/tour/algorithms/data/stack.go
@@ -4,7 +4,7 @@
 // All material is licensed under the Apache License Version 2.0, January 2004
 // http://www.apache.org/licenses/LICENSE-2.0
 
-// A sample that program implements a basic stack.
+// This sample program shows you how to write a basic stack.
 package main
 
 import (
